post-ball-drop training

+++ rl_x/environments/custom_mujoco/robot_locomotion/mjx/environment.py
@@ -136,7 +136,7 @@ class LocomotionEnv:
 
         self.floor_geom_id = mujoco.mj_name2id(self.initial_mj_model, mujoco.mjtObj.mjOBJ_GEOM, "floor")
 
-        self.reward_collision_sphere_geom_ids = jnp.array([geom.id for geom in [self.initial_mj_model.geom(geom_id) for geom_id in range(self.initial_mj_model.ngeom)] if geom.group[0] == 5])
+        self.reward_collision_sphere_geom_ids = jnp.array([geom.id for geom in [self.initial_mj_model.geom(geom_id) for geom_id in range(self.initial_mj_model.ngeom)] if geom.group[0] == 5], dtype=jnp.int32)
 
         self.has_equality_constraints = len(self.initial_mj_model.eq_data) > 0
